+++ library/core/src/main/java/com/google/android/exoplayer2/upstream/cache/CachedContent.java
@@ -29,8 +29,7 @@ import java.util.TreeSet;
 /*package*/ final class CachedContent {
 
   private static final int VERSION_METADATA_INTRODUCED = 2;
-  private static final String EXOPLAYER_METADATA_NAME_PREFIX = "exo_";
-  private static final String METADATA_NAME_LENGTH = EXOPLAYER_METADATA_NAME_PREFIX + "len";
+  private static final int VERSION_MAX = Integer.MAX_VALUE;
 
   /** The cache file id that uniquely identifies the original stream. */
   public final int id;
@@ -94,21 +93,28 @@ import java.util.TreeSet;
     return metadata;
   }
 
-  /** Applies {@code mutations} to the metadata. */
-  public void applyMetadataMutations(ContentMetadataMutations mutations) {
-    this.metadata = new DefaultContentMetadata(metadata, mutations);
+  /**
+   * Applies {@code mutations} to the metadata.
+   *
+   * @return Whether {@code mutations} changed any metadata.
+   */
+  public boolean applyMetadataMutations(ContentMetadataMutations mutations) {
+    DefaultContentMetadata oldMetadata = metadata;
+    metadata = metadata.copyWithMutationsApplied(mutations);
+    return metadata.equals(oldMetadata);
   }
 
   /**
    * Returns the length of the original stream, or {@link C#LENGTH_UNSET} if the length is unknown.
    */
   public long getLength() {
-    return metadata.get(METADATA_NAME_LENGTH, C.LENGTH_UNSET);
+    return metadata.get(ContentMetadata.METADATA_NAME_LENGTH, C.LENGTH_UNSET);
   }
 
   /** Sets the length of the content. */
   public void setLength(long length) {
-    applyMetadataMutations(new ContentMetadataMutations().set(METADATA_NAME_LENGTH, length));
+    applyMetadataMutations(
+        new ContentMetadataMutations().set(ContentMetadata.METADATA_NAME_LENGTH, length));
   }
 
   /** Returns whether the content is locked. */
@@ -234,4 +240,25 @@ import java.util.TreeSet;
     return result;
   }
 
+  @Override
+  public int hashCode() {
+    int result = headerHashCode(VERSION_MAX);
+    result = 31 * result + cachedSpans.hashCode();
+    return result;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    CachedContent that = (CachedContent) o;
+    return id == that.id
+        && key.equals(that.key)
+        && cachedSpans.equals(that.cachedSpans)
+        && metadata.equals(that.metadata);
+  }
 }